Initial commit of forge code and associated data.
This is a first cut at a file structure based on my plans for the code reorganization. Legacy code is preserved in the Archive directory, data, including data collected by the 300H section is stored in the data/undergrad directory with their raw collected data in the raw subdirectory. 

Structure subject to change as the reorganization takes place