continued reorg and initial commit of learning algorithm
organizational upgrades to the main forge code, this involved mainly taking out common code sections and making functions out of them. I still have to implement all of the data processing and output generations from the legacy model, hopefully I will complete that sometime this week. More interestingly I've created the first cut of the learning algorithm. It uses a bag-of-words model for recognition. I didn't go fully into the bayesian probabilities, I'll leave that for another day, but it seems to work pretty well. I experimented a bit with thresholds and different weighting factors ultimately coming up with a pretty simplistic weighting system for words that are directly learned - it's the normalized to [0,1] occurrence in the learned set - and I set up the functionality to use MATLAB built-in genetic algorithms to determine the optimal weighting of the additional factors I added - the actual text of the issues as well as some helper words to work out some of the issues. That code is still running currently, hopefully it will finish in the morning and I'll be able to generate some version of the 3D paraeto frontier. Early returns seem to indicate that the algorithm can match back the learned material at near a 90$ success rate. Probably have to manually code more (or introduce bill-text scanning) to get any higher.