updated learning algorithm
Updated the learning algorithm with some genetic algorithm code to try and optimize the weighting of the additional words and issue words. This involved refactoring some of the script into an easily iterable form. The results are striking but the code needs to be cleaned up to allow for single runs as well as iterative runs. At some point it will likely be easier to drop the genetic algorithm component and do a full factorial search rather than a random number driven search.