finished optimization component
finished and cleaned up the optimization scripts, everything is broken apart into separate functions that are called from the top. Still a bit of work to do in terms of process flow but I think for the most part it is complete. Most of it remains undocumented, I hope to correct that in the next fix. I've also included on this commit the results of the learning algorithm, a spreadsheet that includes the algorithm-assigned issue code and the undergrad issue codes as well as the weights for each different category. The .mat file contains the results of the optimization - this is tied to the specific word list used as well as the threshold (currently zero). Also committing updated figures