finished reorg, updates to look at committees
Fixed the header in forge_legacy so now, if desired, it will run. But  the reorganization is now 95% complete with the remaining 5% being V&V on some strange results from my initial recalculation. Things are now set up to use the bill structure and it's relatively quick running. More work needs to be done on optimization and abstraction. Currently everything is focused on the House and (though we're not looking at the Senate yet) will not work for the Senate. There's significant strangeness happening in the sponsorship data and in the committee data, there are several things that need to be worked out before it can go live. But it is a rough cut and it appears that the main vote results are identical. Included in this commit are updates to some of the data collected by the undergrads, initial commits of the basic starting data for running forge without compiling, and the commit of results including the gifs, histograms, and surface plots