further updates
Upon reflection it made more sense to just break everything out into separate files so that's what I did. Slight updates to some of the data sets, added a new copy of the learning_algorithm_data.mat file to the data directory, this is the one that will be used by the main forge algorithm, it can constantly be update without having to deal with the date string.