validation and updates to the core forge model
Okay, this is the big update. I fixed (as much as I could) the issues surrounding the sponsorship and committees. There's a small number of bills that are competitive and an even smaller number that we have committee data for (part of that seems to be time-related, Legiscan doesn't have reliable committee data for the older sessions). All of the outputs are now generated and saved out, I've regenerated the histograms and everything should be hunky-dory. The next step is to take these outputs and actually do the model with Bayesian updating.