bayesian update
This is the big one. I *think* I have successfully implemented a low-level bayseian updating scheme based on sponsorship and committee votes. I still need to work out how to tie in new information ad do the updating (and then generating the all important pretty plots).And then of course the optimization because right now it's incredibly slow. Along the way I fixed several things in the main code, in particular identifying bills that are competitive. Now it is limited to competitive bills that have data for both a committee and the main chamber. It also correctly parses the third reading rather than reconciliation bills. This leaves a total of 23 bills. A little disappointing but more than enough for a proof of concept set. The next step will be to generate the bill histories for each and allow the undergrads to select the one they want to research. I don't expect that will be too difficult. Pretty pleased with the way things turned out, even in its current state at t_2 the bayeseian updating is predicting with better than 60% accuracy, averaging just over 80%