bill information update
Updated the output for the competitive bill list to include the names of sponsors and committee members. Should make it easier to do a legislative history involving new legislators revealing preferences. Slight modifications to the broader forge code to create histograms of the sponsor and committee membership for the competitive bills