compatibility with other states
Some modest reorganization to accommodate expansion to multiple states. Outputs are now saved in state-specific folders. Turns out there are only a small number of changes necessary for basic compatibility. The majority of the code is generic enough that it can be applied between states. In the test with Ohio there was some significant issues with separating out the committee votes. It's also necessary to separate out the different parties and work out the specific names of the legislators for the session of interest. Current implementation is basically just a proof-of proof-of concept. But it works! That's kind of crazy