full bayseian prediction
Put in the basic implementation of logic to do updating based on revealed preferences. I've done it in such a way that should be extendable to any number of revealed preferences for any specific bill. Eventually this will be done dynamically as the crawler will be able to search for revealed preferences but for now it's done by reading in a spreadsheet that has the date of each new revealed preference, the legislator responsible, and the direction of the preference. Early returns are really good - incorporating the revealed preferences for 15 legislators brings the predictive accuracy to 97$