accuracy update and seat matrix calculation
Revised the accuracy calculation to do updates with each new t_n. The spreadsheet also now includes legislator names for easy analysis. Reincorporated the seat matrix calculation into the general calculation. In its current form it will break if applied to states where we don't have seat matrix information. As part of this it was necessary to recompile everything thus the massive commit