core forge updates and new bill prediction
Updated the core forge system to allow for randomized bill preference selection to be used in the bayesian prediction process. This is currently limited to the bills under consideration by the undergrads but there's not reason it couldn't be used to predict any bill meeting the basic criteria of committee votes and chamber votes. I also added a function to convert string ids into numeric ids, just to make life easier.