refactor update
I'm seemingly always refactoring things. I've broken out some of the basic functions into separate utilities, the +plot work is basically done, +predict is not - there are still dependencies on the basic forge code. Perhaps more importantly I've broken out the components that are state specific into a separate "IN" function. The basic development plan from here on out is to make a separate, subclass of forge for each new state. Things that are relevant to all states can go up into the forge super-class. More work to be done!