update and reorg to include 2015 and 2016
Fixed some things in the core forge and IN methods to correct a location issue for state-specific data. I also moved around some of the functions in the predict function and reconciled an issue in loadLearnedMaterials related to the state-specific reorg. I uploaded the 2015 and 2016 datasets for Indiana and did a full recompile for all plots and charts.