massive restructure
I updated the read-in of the people data to allow for the more complete data provided by legiscan for the 2015 and 2016 sessions. In doing so I made it easier to access people in both the house and senate and, as a result, updated the core forge model and associated IN model to calculate all of the relational matrices for the Senate as well. As is becoming tradition this is a full update to account for those changes. The previous commit cleaned out all the outputs, this commit puts them back though it is likely I will move things around to reduce clutter.