Formatting, monte carlo, and recompute
Formatting changes across the board, this was an in-progress optimization run to try and cut down on the simulation run time during testing. It is not a substitute for a more in depth revamp but it's sufficient for now. I began the process of breaking the prediction algorithm into a separate file, that's the revised "updateBayes" function. Additional functionality was added into createIDstrings to do an ismember check before leaving the function and createTable to return multiple tables of the same type, cutting down on function calls (theoretically). THe biggest part of this commit is the first cut at the monte carlo simulation of results for the predictive model. It needs cleaned up and could probably use a lot more work but it's there. The initial results are good, they have to be run again so I guess this is technically a WIP commit.