Completed monte carlo
Some small optimizations and improvements in advance of the monte carlo predictive runs. I'm fairly convinced that's the next big reorganization - to speed things up drastically. This will have long lasting impacts and is really what needs to happen next. Included in this commit are graphics for both the house and senate for the accuracy generated as well as the change between t2 and tn (the detla)