full recompute and update
As is becoming tradition when there's any sort of substantive change I've done a full recompute (again). I'm thinking about eliminating the outputs from the repository because this is so tedious to have to do every time. That's a question for another day I suppose. Also in this update I've started tracking profiler runs, this will allow me to check on speed differences a sort of "nightly run" esque procedure. I also made a small change in startup to exclude the Archive and the Profiler directories from the MATLAB path. The next steps will be to do a optimization and performance review which will involve changes and profiler runs.