massive performance improvement
With a few more changes I've managed to bring runtime down from 6700 seconds to 59.8, a better than 99% reduction in runtime. Not sure if it means I'm good at speeding things up or bad at writing code the first time. Regardless, this is now in good shape to do larger/more complete runs to get a better sense of pivotal legislators.