recompile and recompute
Full recompile and recompute given the recent updates for optimization in one part of the code. I am also including the profiler results from the full sweep - there aren't a ton of improvement areas, it is pretty clear that output generation is a big drain on the runtime (no surprise there) but there's not really a good way to make it go faster other than not running that portion. All of this is in preparation for the full monte carlo analysis of all of the bills which will generate the dataset I need to start examining pivotal legislators.