further updates to prediction model
Even more updates, the runtime has gone up and down a little, I've made some significant performance improvements across the board to cut down on runtime and increase the common code base. I have yet to test a few of them so there will likely be a patch update in the near future. That said, things are humming along nicely and I now have the full results for all of the house and senate bills. The next step for this will be to start analyzing the critical legislators. It is possible I haven't collected enough information about the model accuracy in order to determine the pivotal legislator, a fix that would drastically increase runtime again.