Prediction function updates
Updated the prediction functions to track individual changes in accuracy on a per-legislator/revealed preference basis. I think I might need to include a starting accuracy as a weighting for the function but I'm not entirely sure how that's going to go. At this point I'm just happy to have the data laid out. I also fixed an embarassing naming issue, these are boxplots not histograms. And I generated boxplots for the total performance on a per chamber basis.