critical legislator update
So I created a basic formula for determining critical legislators. I gave them a score based on the percent change of the t2 accuracy of the delta of their revealed preference. So the t2 accuracy (t2_accuracy) and then the change in accuracy with their revealed preference (delta = t_n - t_n-1). Times the weighted average of the position of their revealed preference, the first position being weighted the highest. I'm not sure I like the resulting metric but it seems to work okay. I discovered that not every legislator is getting coverage over 100 montecarlo runs on a per-bill basis so I'm going to up the number of runs to increase the available data. That will be next. Given the runtime optimizations I am planning on doing a monte-carlo number of 1000, perhaps higher.