big cleanup and upgrade
Clean up of almost every function with comments and updated spacing. I also reworked a few different things to remove depreciated or defunct code and make it into a more accessible format. The biggest thing is I've started creating more states - Ohio and California are now in the beginning stages of being supported. I realize I need to create a new class which I'll probably call "state" that each state will inherit from. This will contain the general run functions that are abstractable to each state. That will be a future commit. As part of this update I've regenerated a bunch of plots and generated new ones for the new states. They're using my other computer so the color scheme is (annoyingly) different, that will be fixed in a future run I'm sure. The important thing about this update, potentially the most important thing I've done in the last few weeks, is now there is the functionality to test on multiple states. It's going to make testing a little more difficult which is why I've done an initial commit of the script "tester" which is the start of a unified testbed for updates. More on that to come.