data and output update
Follow on to the previous commit (just seemed to make sense to break it, no real reason to do it) these are all of the legiscan data files for CA and updated files for OH. And a full run of all three, at least to the breaking point. Currently Ohio is wonky, there's an issue identifying the final vote as the legiscan data doesn't identify Third Readings in the Ohio set. Likely this means adding a flag to sort by date and then take the last one (ugh). California does identify the third reading but the prediction analysis identified more than a thousand competitive bills in the lower chamber. In contrast there were 66 in the upper chamber. This is likely a bug and will need to be fixed. Also the color scheme issue.