Cleaup and revised testbed
Moved the forge init call to be in the state run method - this allows for easier testing in the testbed and doesn't affect functionality elsewhere. Also updated the testbed to be a more robust script. Still not complete but should be a good start in helping work though some of the strangeness that has been happening