bayseian updating fix
Fixed issue with the implementation of bayseian updating. Basically I had make some changes to speed up the operation and I messed up the math. The result was greatly exaggerated performance. THe fix is now implemented across the board. I also changed the way in which legislator impacts are calculated. The procedures for these are being compiled into a user manual which will be uploaded soon.