restructure and revamp
First step in restructuring the forge data to be more accessible. Changed the naming convention to be more consistent and make more sense (finally); restructured the monte carlo analysis to do the full prediction up front. As of yet it's only been done with a *very* small monte carlo run (10) for both IN and CA. Given the issues with OH it's been dropped from the testing rotation.