prediction model updates (final)
A few changes to try and optimize the prediction model. This separates out what will be the very large monte carlo data in order to minimize the file size in the outputs directory, reorganizes the output to identify the monte carlo number. This is an in-progress commit and will be followed up with some big-hitter commits with the compiled data files for 10K, 50K and maybe even 100K monte carlo numbers