ELO Prediction first cut
Completed the first version of the ELO prediction with two different K weighting schemes - one variable (though the rates likely need adjusting) and one fixed. In the process of doing this I've uncovered a few issues with the legiscan data and a few shortcomings with the way I'm handling the legiscan data. What I'll do next is a reorganization to put all the legiscan data together in one place, rather than distributed to all the state output folders, and change around the state information data from the separate classes of the current instantiation to a "settings file" model which should accommodate more rapid expansion of the program to other states. This is a full recompute of IN with a MC# of 1000.