legiscan data update
Moved all the legiscan data into a single "legiscan_data" directory. Also added in data from the remaining 47 states, the District of Columbia, and the United States to play with. this should make it much easier to keep track of all the state data though it is still necessary to go through and check that each state is formatted and parsed correctly.