IN PROGRESS COMMIT
Okay, I've settled on a path forward for this reorganization and a new test case: the US House and Senate. In messing around with the US it's become clear that the committee data is unreliable at best or (in the case of the US) non-existent. So I'm going to strip out all of the committee analysis stuff and remake the prediction models to just start with sponsorship, leaving out committee votes and consistency. This is a *major* change that's going to result in a whole hell of a lot of changes. Some of the files being committed from this initial US data-output commit will be erased as a result. All of the monte-carlo analyses will have to be re-run and for more states. 

Other things that are added in this commit:
-support for states that don't have clearly identified third readings using the words "Third" or "3RD", added support in for "On passage" which is the identifier used for US data
-created the state_properties file which is called by a revamped state init
-cleaned up the way folders are created that should allow for a smoother process - all folders are checked for existence and created by default at initialization