Getting closer
Worked out the bugs in implementing the bill classification into the state code. Next step will be to clean up the learning algorithm ops, probably a temp folder in +la, and generation of basic statistics about the learning set. From there it's resurrecting ELO, implementing it for different bill categories (with a minimum threshold), and implementing SNEBM