Concise Update
In an attempt to cut down on accuracy issues I've manually cut the number of categories from 36 to 12 in what I'm calling "concise codes" - I'm terrible at naming things. This is an in progress to serve as a log. I've recompiled the learning table with the concise data (lots of flags, eventually I might want to put an additional check in there to see if it's concise). I'm going to run the big algorithm now for the first time and see if we do any better.