Mess cleanup
Continued cleanup of all my messes. Eventually all of this will get deleted when I settle on a final version but I want to try and maintain things so I can revert back if need be. Biggest thing: I've implemented and reimplemented the concise categories, rather than 32 issue categories they're lumped into a cool 11. This seems to work a lot better (accuracy is up above 65%!). I rejiggered the categories after the last big optimization run so I'm going to optimize again. I'm also going to quickly test to see if the ELO code works correctly and, if so, run that again to get some quick sanity-check outputs before kicking off a larger analysis. I think the LA stuff is nearly in its final state, if this optimization goes well I'm going to put a bow on it and call it done.