ELO really fixed this time
ELO is up and running. The issue was I had depreciated the committee analysis in other parts of the code but that hadn't filtered its way down to the ELO model. But that's all updated, I've done another round of outputs at the state level. Everything looks good to go for an in-depth analysis run