Fixed the ELO issues (I think)
So there was some issues surrounding the ELO model in the large run. For one, I forgot to turn it on. But I've gone through and fixed it up, I think it should be all set now. This is the code commit, I'm going to do a clean wipe of the data branches to clean out all the old and unused stuff, the next commit will be an update of all the data and a delete of all the unused material. Additionally, rather than doing tons of montecarlo runs I've settled on 2500 which should run relatively quickly and still give me the correct spread. I noticed some potential issues there and will be looking for them to manifest in this next run.

Next steps for the ELO model: adding in bill specific comparisons