ELO Update
I did it! ELO is now monte carlo-able, this is the initial commit of that capability. More to come shortly, I am about to do an optimization test and should have results from the first mc1000 run later today. My plan is to run a 10,000 (or maybe even more) run over the next week