First ELO MC Commit
Completed a run of 1000 seeds, this is the first push. It was generated using the code from before the last commit so it took over 20 hours to run. With the optimizations in the previous commit it should be able to do it in half the time. The next run will be a long run of 100K seeds