The Sunday-Funday Update
I re-added the Undergrad Data which had gotten dropped in an earlier commit. Added in some lines to bring that data into the Elo set and combined it into the Elo data. But because of issues I discovered in the Elo code I'm going to rerun it with a large monte carlo set