Seniority Data initial commit
Initial commit of the code (Stata!) to parse the seniority data out of Chris's data set. The do file currently only keeps things in that local directory, eventually I'll make it so it puts the data file in the right place. First I have to figure out where the right place is. And ultimately this may be overcome by just cutting the finance data in directly from stata rather than the excel pass through that the current version was done with. Anyways. Not committing the full finance data scripts to protect Chris's data so this do file is pretty useless. Outputs are solid though!