Graphics Update!
Some changes to allow generating relational matrix images for each bill category. It led to some interesting discussion for possible future work, summarized below. Food for thought!

The other thing to keep in the back of our heads (not for Southern but maybe afterwards) is a process question. Right now we generate a relational matrix based on the full billset (say 100 bills) and for each category we run that full relational matrix on a subset of bills. We may see different results running a relational matrix generated from that subset. We could also (separately) run statistical analysis on the different relational matricies.